divisonCalcv.1

+++ programming_paradigm/robust_division_calculator.py
@@ -1,6 +1,6 @@
 def safe_divide(numerator, denominator):
     try:
-        result = numerator / denominator
+        result = float(numerator / denominator)
         print(f"The result of the division is {result}")
 
     except ZeroDivisionError:
